requirements.txt
aiofiles==0.8.0
anyio==3.5.0
asgiref==3.5.0
CacheControl==0.12.11
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
configparser==5.2.0
cryptography==37.0.4
cycler==0.11.0
DateTime==4.4
deprecation==2.1.0
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
EasyProcess==1.1
entrypoint2==1.0
fastapi==0.75.2
fonttools==4.29.1
h11==0.13.0
idna==3.3
joblib==1.1.0
jsonpickle==0.9.6
kiwisolver==1.3.2
matplotlib==3.5.1
MouseInfo==0.1.3
msgpack==1.0.3
mss==6.1.0
Naked==0.1.31
numpy==1.22.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.0
Pillow==9.0.1
psycopg2==2.9.3
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.0
PyGetWindow==0.0.9
PyJWT==2.3.0
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2021.3
PyYAML==6.0
requests==2.27.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
shellescape==3.8.1
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
squareup==11.0.0.20210513
starlette==0.17.1
stripe==3.3.0
threadpoolctl==3.1.0
typing_extensions==4.2.0
tzdata==2021.5
urllib3==1.26.9
uvicorn==0.17.6
xmltodict==0.13.0
zope.interface==5.4.0
